refactor function calling logic to ClassificationTemplate
